[zircon][usb-virtual-bus] Add integration tests for the UMS role
This patch adds integration tests for the USB Mass Storage peripheral on the USB Virtual Bus.

The integration test shells out to the standard blktest executable, then tests
connecting and disconnecting the device with and without write caching on the simulated device.
Devices with a write cache will lose data that is written to them.

Change-Id: I5b760a791c1e3b88e86881875b214df8f4e7db2a