[usb-device] Don't assume order of callbacks.
Callbacks may sometimes complete out of order, even if successful e.g. in ZX-3542.
This can happen if a request is completed without being queued onto a
transfer ring (e.g. callback due to queue error).

We do an immediate callback if it is requested / completes out of order
/ has an error.

If the request completed in order, we also return the number of silently
completed requests back to the client, so they can verify.

ZX-3542 #done
ZX-932 #comment

TEST= ran usb-test, checked usb-audio still works

Change-Id: I7340e71b8ece37cfa27fd6bea59bffe38f31f74d