[zircon][hisi-ufs] Add LUNs as Block device and support to run lsblk
Currently this driver probes every LUN in UFS device, Read INQUIRY,
READ capacity and add them as Block device. The basic READ operation
is implemented. The UFS LUN block devices are created in
/dev/class/block/ and lsblk lists the block devices and their partitons.

TODO: The WRITE operation is not supported. Also support for
parttioning, mounting, read and write to the disk, to be added.

Test: lsblk lists the UFS LUN block devices and its partitions.
$ lsblk
ID  SIZE TYPE LABEL FLAGS  DEVICE
000   3M               /dev/sys/platform/02:01:5/hisi-ufs/ufs-disk-0/block
001   3M               /dev/sys/platform/02:01:5/hisi-ufs/ufs-disk-1/block
002   7M               /dev/sys/platform/02:01:5/hisi-ufs/ufs-disk-2/block
003  29G               /dev/sys/platform/02:01:5/hisi-ufs/ufs-disk-3/block
004  29G               /dev/sys/platform/02:01:5/hisi-ufs/ufs-disk-3/block/part-000/block

Change-Id: I7c13e7e923ad4e38fedc9df5dabdb223d60ee403