[fs-test-utils][blobfs] move bloblist to fs-test-utils
the blobfs integration tests had a useful structure for reading,
writing, and keeping track of random data blobs that is going to be used
in the power failure tests. this pulls it out into the fs-test-utils
package and converts it from a struct into a class. it also replaces the
use of the original functions with the new class in the blobfs
integration tests.

Test: /boot/test/fs/blobfs-test
Change-Id: I5d839bc51c5620a3414e0bd0ae83070f6c7d70ac