[zircon][ahci] Handle timeouts better.
If the watchdog timer fires and completes a transaction, the IRQ thread
may not be aware and assert on an unexpected hardware operation.
Set more state bits in the watchdog, and make the interrupt thread aware
of them to avoid this.

Add logic in watchdog timer to identify if we're getting spurrious
timeouts.

Tests: blktest and fs-test on ahci
Bugs: ZX-3548
Change-Id: Ia26fe696617d442d6e835d2940b9b72cba085e2b