[media][tests][e2e] VAD notification frequency API
This CL adds an API to the fuchsia.virtualaudio.Input and
fuchsia.virtualaudio.Output protocols, enabling VAD clients
to override the cadence of position notifications received.
AudioCore sets the number of notifications per ring buffer,
when the ring buffer is established, and currently sets a
value of zero for all streams (no position notifications).
Thus for virtualaudio clients to receive position
notifications for an individual input or output, they must
override the value that AudioCore sets, with this API.
This API can be called before the ring buffer is
established, although it may be more useful to wait until
the buffer size is known, so that notifications can be set
to a time-based cadence.

Note: this API does not cause VAD to generate "official"
position notifications to AudioCore -- but rather to emit
equivalent notifications that go only to VAD clients.

This CL also includes updates to the virtual_audio utility,
as needed to validate this work.

Although an integral part of our end-to-end testing effort,
the virtual audio FIDL interface, driver, utility and
service are test-only. This CL contains no product changes.

To observe these notifications manually from the shell:

 # As a prerequisite, Activate AudioCore and virtual audio.
 vol --show && virtual_audio --enable

 # By default, AudioCore requests no notifications.
 virtual_audio --add --in --add --wait

 # The 'notifs' flag overrides, here setting 2 notifs/ring.
 virtual_audio --notifs=2 --add --wait

 # If this flag specifies no value, it is set to 4.
 virtual_audio --in --notifs --add --wait

 # Configuring notifs can be done before or after the Add...
 virtual_audio --add --notifs=5 --wait

 # ...or both! It can be changed while a device is active.
 virtual_audio --add --notifs --wait --notifs=1 --wait

 # The value zero (no notifications) is also valid.
 virtual_audio --notifs --add --wait --notifs=0 --wait

 # As mentioned, actual frequency is linked to buffer size.
 # This configures an output to 12 notifs in a 6-sec buffer
 # and an input to 2 notifs in a 1-sec buffer (equivalent).
 virtual_audio --notifs=12 --rb=2 --add --in --notifs=2 --rb --add --wait

Test: build, CQ, manual verification (described above)

Change-Id: I255e94149b1475584f7b4d563c16507d2c7c03d7