[debugger] Move (some) command line switches to options.
Moved over the ones that represents configurations. This is a first step
to create a configuration file that can be loaded into zxdb.

This is necessary now that repo handling is being made by switches, so a
way to use files to configure zxdb becomes more important.

Simplified a bit quitting the agent.

TEST=manual

Change-Id: I355e46b5376343d5ceb80d32e44a678a3b0f7a5c