[debugger] Better canonicalization of minidump UUIDs
Minidump packs build IDs into its UUID field. The string conversion of
this function leaves some hyphens in it which we don't need, and also
zero-pads them.

Furthermore, there are endianness issues with UUID encoding because it
is at least partially a stream of integer values rather than a stream of
bytes. There's a lot of quirks involving this in our pipeline today. For
now we add an endian transform that seems to result in us matching
readelf end-to-end.

There is an additional problem that some build IDs are 160 bits, and
have to be truncated to fit in a 128-bit UUID. We'll have to work around
this in another patch.

Change-Id: I1e5668a58140337fb7e77c468f5457b2c48731f8