[debugger] Add untruncation logic to build ID index
We occasionally encounter debug IDs that are truncated to 32 characters
(16 bytes/128 bits). The build ID index should be able to recover the
last characters for us and match anyway. The longest build IDs we expect
are 40 characters long today, and of the formats in our tree, the newest
is actually shorter.

Change-Id: I148fe446a4310a75c54d14160f2ca1edda270317