[run-test-component] resolve uris before loading
This change isn't quite the ideal change of it's nature, it would be
preferrable to use the fuchsia.pkg.Resolver API instead, but that is pending
more changes that would potentially block things. Instead using
fuchsia.process.Resolver should be sufficiently stable for now, and provides
the same side effects. This change ensures that manual invocations of
run-test-component always use the up to date CMX file to setup environments
around tests, rather than using the previous environment, as before this
change.

Change-Id: I29536a5b83ae360c65d36d84871109b13d1a6773