[ledger] Don't use GetObjectPart when getting the whole object.
GetObjectPart was introduced in CL 232012 and GetObject was converted to
use it instead of duplicating the code. Unfortunately, that meant that
for small objects we would do the conversion from string to VMO and then
back, which had a noticeable performance impact.
This change brings back GetObject to (roughly) how it looked before CL
232012 + addresses one other issue with that change (missing return;).

TEST="fx run-test ledger_tests" still pass; get_small_entry benchmark is
significantly faster when run locally (0.29ms -> 0.18ms).
LE-689 #done

Change-Id: I15f03b75dae748f2843ad3cd5737788462b0e4ac