[debugger] Activate async message loop as default.
Enabled a flaky test about debugged jobs.

A previous change moved restarting a newly created process only if it
was launched by a filter (used for launching components). This is wrong
as it has to be done regardless. Moved restarting the process back to
the process starting event.

Added some more logging statements.

TEST=manual

Change-Id: Id59862f561b62377fc6b3d64e781eec338242980