[banjo] Add support for parsing fidl IR JSON into structs.
This lands initial support for parsing fidl IR JSON. The next step will
be to take the deserialized structs and merge it into our AST
representation so that the types can be used as easily as banjo types.

The JSON schema doesn't match up entirely with the output from fidlc, so
the representations doesn't line up with the schema perfectly.

The test is not comprehensive as it is missing many variations of
structs.

Tested: fx run-host-tests banjo_host_tests
Change-Id: Id8603e829a876ddfdfa09c755fd085403b897854