[pkg_resolver] Return ZX_NOT_FOUND or ZX_UNAVAILABLE if repo is down
This extends the package resolver to return if a package is
not found in the tuf repository, or if the the tuf repository
or blobs in the package are currently unavailable, instead of
just returning an internal error.

This also rewrites the resolver tests in order to allow us to
create synthetic test files, to make it easier to simulate
certain error conditions, like a package being unavailable.
It also switches over to using `fuchsia_sync::run_until_stalled`,
which will error out the test if all the futures get stuck.

Test: Added tests to verify resolver is able to report missing
or unavailable packages
Bug: PKG-607 #comment better error messages
Bug: PKG-520 #comment better error messages

Change-Id: I92bb16bae9a795be10d836cf16ab234a0903d3e9