[auth] Add option to force token deletion if revocation fails.
Recently we've been seeing some problems with deleting accounts while
the network is unavailable following a "factory reset". Originally this was
because the deletion was being performed right after a reboot before the
network came up. At the time SysUI and Identity agreed to move the
deletion as a stopgap and add a force option as a longer term solution.

Unfortunately last week we made a different change. This CL implentments
the longer term solution we'd previously agreed, letting clients choose
whether its more important to maintain consistency between client and
server state or to actually get rid of the tokens.

Change-Id: Idc3c12028ad4d6a19cf3d7f81b6a54eeda4a611a