[bt][tests] Enforce ordering of device creation and teardown in tests
Fix a test flake in the Bluetooth integration tests caused by undefined
ordering of behaviors and state in reasoning about available hci and
host devices. This caused some expectations about the availability of
the fake devices under test to be falsified.

We now use the Control interface to wait for specific states of bt-gap,
and ensure that the bt-host addition and removal are strongly ordered
with respect to our test behavior.

This commit also improve failure messages for test predicates.

FLK-106 #done

test: runtests -t bt-host-integration-tests

Change-Id: I7a5278414639b00698dd57d2e3063f5f941f67c7