[pseudo-fs] Split directory::connection from directory::simple.
Other kinds of directories will be added in the future.  The first one
will be a lazy directory.  They all will need to deal with the FIDL
connections.  So, extracting common functionality into a separate file.

Simple directories order their entries alphabetically, but a lazy
directory may use a different kind of ordering.  Making
DirectoryConnection polymorphic on the traversal position type will
allow it to be used with different kinds of directories.

Change-Id: I5571c847aa467f9f08cf4691cfffa22d93a56c4d