[zircon][fdio] Retry partial writes
The Android Java debugger was trying to send a packet that was too big
for the socket buffer. This did a partial write, which the debugger
interpreted as an error and closed the connection. On Linux, a blocking
write to a socket will write everything (unless interrupted by a
signal, which is not a concern on Fuchsia).

There is no need to update the implementation of read: on Linux,
attempting to read more than is available from a socket will not block,
even if the read is blocking. This matches the current behavior of fdio.

Test: Linked Android runtime with modified fdio, attached debugger. Also
ran included unit test case.

Change-Id: If70db478ed4034920e2ef60ee1bcd6866c872698