[fidlgen] Requiring explicit paths in test utilities
With the re-use of the test utility for fildgen_dart, we need to provide
explicit paths since the JSON IR goldens are in one place, and the
generated code goldens are in another. This will also be true when we
split fildgen into separate backend per target binding e.g. fidlgen_cpp,
fildgen_rust, or fidlgen_llcpp.

Test: fx run-host-tests fidlgen_{cpp,llcpp,rust,golang}_test
Change-Id: I8ea0f635f02113344ba7e41b4818866249b44534