[sdk] constrain type for InterfaceHandle(type)
Previously, this templated non-explicit constructor would try to act as
a type conversion from any type T to InterfaceHandle.

With this change, use SFINAE to prevent the compiler from considering
this constructor for implicit type conversion unless the source type is
at least roughly suitable, and static_assert() with a message if the
source type isn't exactly one of the intended specific types.

With this change, an attempt to pass a type whose remove_reference<>
type isn't one of the intended types will ignore this constructor and
complain that the provided type can't convert to the needed
InterfaceHandle (good), instead of (hopefully) complaining that the
provided type doesn't implement Unbind() (not ideal) or (worse)
accidentally compiling if the provided type happens to implement
Unbind() despite not being one of the intended types (potentially bad).

The intended types (based on existing usage) are InterfacePtr<Interface>
or SynchronousInterfacePtr<Interface>.  For now, we don't permit
subclasses (is_same currently) because there are no such usages, but
in future we could allow subclasses (using is_base_of) if needed.

Tested: CQ
Change-Id: Ib49c4a431070ed12db5104007e2d25fdb5ba0d38