[zbi] Move BOOTFS format decls into <zircon/boot/bootfs.h>
Also update the spec comments for ZBI_TYPE_STORAGE_* to clarify its
role in the ABIs between boot loaders and the rest of the system.
These items are always opaque to boot loaders, so they need not have
strong compatibility.  A given ZBI is packed with the kernel and
core services that go with the storage-typed items and their formats.

Change-Id: Ib884657b970ba89654946aa591b1af72bdd814b6