[mediatek][display] bug fix related to overlay engine stalling
The display driver was crashing sometimes when trying to apply a new
configuration and the overlay engine (OVL)  would simply not stop. The issue is not
easily reproducible and requires 7+ hours of running "display-test
--mediatek" in constant loop until it happens. The root cause of the
issue was most likey due to the fact that the MUTEX engine was not
stopped before the OVL engine was stopped and therefore we might have
hit a situation where the MUTEX engine is triggering an event to the
OVL engine and since the OVL engine was being reset/stopped, it was
causing it to go into a non-recoverable state.

Test: Tested on mediatek hardware and ran display-test --mediatek in
constant loop.

Change-Id: I8dd0b319398f13dcc9e5b0401868057661088810