[buildidtool] Introduce tool to manage .build-id
This tool uses stamp files and dep files to make .build-id entries a
first class citizen of the build system. Each stamp file acts as a
place holder fot the associated .build-id entry. The tool is carefully
designed to be atomic and work and many bizzare rebuild senerios. There
are some invalid states that you can get into but they involve leaving
additional files on the disk and do not invalidate your build. When the
tool functions normally, when it overwrites the stamp file it also
overwrites the .build-id entry that it was previouslly associated with
to mimic the .build-id entry being overwriten. A dep file is also emited
which means that if a .build-id file is altered or deleted it will be
rebuilt (this is somewhat counterintuitive)

Test: I implemented the GN changes for this and built Fuchsia under many senerios + CQ
Change-Id: Ia761952761c1d299b41fbc72022b44b8b6e3bedf