[futex][test] Fix FLK-129 and re-enable futex-ownership tests.
See FLK-129 for details.  TL;DR, make sure that when sharing variable
between threads which are not protected with locks, that atomics are
used in order to ensure memory coherency.  This is particulaly
imporant on really big SMP machines where this isn't just automatic.

As extra insurance, upgrade all of our atomic accesses to sequential.
This is clearly more than is needed, but it is more imporant for this
test to be non-flaky than fast.

FLK-129  #done
Tests: Multiple runs on Cavium bots.

Change-Id: I68d90ee843bc69242df047ea96ef3c6e00123ea7