[build-zircon.sh] add -z and temporarily deprecate -o
-z is for specifying the zircon build out directory while -o is only for
the parent out directory.

This is the first change for a multi-step soft transition.

1. This change
2. Migrate the CI recipe to -z
3. Make -o equivalent to -z
4. Move the recipe back to -o

This is the first step of making it possible to select the compiler for
building zircon. But it seems there are many places where the name
"build-zircon" is hard-coded. I'll see how far I can go with them.

Bug: BLD-383 #comment

Change-Id: Ibf90151a31e75eb301593f73d6a8e2eee467dcd1