[mediaplayer] create mediaplayer stand-in
This CL copies all of the mediaplayer source code and creates a new
version of the mediplayer binary/package named mediaplayer_tmp.
mediaplayer_tmp is registered as the implementation of
fuchsia.mediaplayer.Player.

This was done as part of a soft transition in which the
fuchsia.mediaplayer library will be renamed fuchsia.media.playback.
mediaplayer_tmp will prevent out-of-repo clients from breaking
while the original mediaplayer is modified to implement the new
interfaces.

TEST: no behavior change

Change-Id: I9c0eef006175888a0159db84205427edebc439f7