[perfcompare] Implement comparing results from performance tests
This implements the comparison using confidence intervals described in
the design doc.

Initial testing indicates this doesn't yet work very well when tested
against the perftest-based tests on NUC bots: It outputs "faster" or
"slower" often (that is, at probabilities greater than ALPHA) for
pairs of results from the same Fuchsia revision.  This is probably
because times for test runs are not IID, probably because there is CPU
or OS state that causes times of adjacent test runs to be correlated.
This will require some further work to address.

Ideally we would use the t-distribution/t-test for calculating the
confidence intervals and use SciPy for that, but to avoid that
dependency for now we use the z-test (normal distribution) which is
easier to do.  For perftest's current sample size of 1000
(i.e. running each test 1000 times) this doesn't make much difference,
and the problem above of results being non-IID is a bigger problem.

Bug: IN-646
Test: "python perfcompare_test.py" + perfcompare trybot
Change-Id: I21f4cb877a49e463325019ac8b2c336f86d0ee94