[handle] fix KernelHandle performance regression
The initial KernelHandle implementation made one additional RefPtr copy
compared to the raw RefPtr code, which turns out to be a significant
regression for quicker objects (5-10% for EventDispatcher creation).

This CL reworks the implementation slightly to move the entire
KernelHandle into Handle creation, that way we don't have to copy the
underlying RefPtr and it still gets torn down properly if Handle
creation fails.

This fits better with the established Handle API anyway, it probably
should have been this way from the start just based on code style.

ZX-3800

Test: k ut handle
Test: fx shell /pkgfs/packages/garnet_benchmarks/0/test/sys/perf-test \
          -p --filter HandleCreate_Event$ --runs 1000000 (NUC)
Change-Id: Ic2cc5cb6a87fd03bb9fc78c40a75ba1eafa88390