[zircon][mbr] Fix race condition unbinding MBR
MBR driver's bind() call launches a thread to read the disk and add
devices for each partition it finds. There is a race between this thread
and the unbind() call, which may unbind the first partition found before
this thread has done its work. Added a completion indicator to
mbrpart_device_t that unbind() can block on before deleting it.

Bug: ZX-3700
Test: Bind and unbind MBR partitions.
Change-Id: Icfb6b817dfa9ac5d8f96ae4fb1a2fcebff6a4a6b