[runtests] Add -r option to repeat tests
Unlike gtest, dart test does not have an option for repeating the tests.
Furthermore, some flakes may manifest in the setup of hermetic test
environments (those due to starting injected services).

Also adds a flush to the RUNNING TEST banner, which was not consistently
flushing for me over fx run-test, as well as a progress banner for all but the
last iteration.

Test: fx run-test driver_example_mod_tests
      fx run-test slider_mod_tests -r 100
      fx shell runtests
      fx shell runtests -r 3
Change-Id: Id5cb4a548420b591b989db9669dc97ad94b635d9