[omaha-client] Common and configuration modules
This is the initial check-in of the omaha_client::common and
omaha_client::configuration modules, including a number of types that
are necessary for defining the data that's needed to generate a
protocol::request::Request.  This is support work for implementing a
RequestBuilder that returns a protocol::request::Request based on types
that are more easily managed by the client.

Bug: PKG-468 #initial check-in of support types
Test: fx run-test omaha_client_lib_tests
Change-Id: I375cefdae92c5300abb96058fbd865054b4084bf