[media][drm] Remove limit on InitData size
The InitData structure size vary according to their type (CENC, KEYIDS,
WEBM, etc). The 16 byte limit that was originally in place was actually
intended for init vectors for decryptors, not the init data. Since we're
not limiting the InitData types to a pre-defined set, it is difficult to
have a known limit for the size of the init data payload. The byte limit
for init vectors will be added in a subsequent CL.

Test: Compiles. There are no consumers of this API yet, so there will be
no breakages.

Change-Id: Ie6248f7d3b89a2f6d9d7d2561e9448773ae7f040