[omaha-client] RequestBuilder initial check-in
This is the RequestBuilder's initial implementation.  This commit adds
the RequestBuilder itself, along with supporting structs.  With this,
the omaha client can declare all of the pieces: configuration,
applications, and then use the builder to construct an object that has
all the right fields for making a request to the Omaha service.

Bug: PKG-467 # RequestBuilder itself
Test: fx run-test omaha_client_lib_tests
Change-Id: I80890ac957fd7648b8b97184a1f11e65560e09ee