[pseudo-fs] Extract DirectoryConnection::connect().
Move some of the connection initialization logic from directory::simple,
into DirectoryConnection as it will be shared between different
directory types.  Plus this logic is required for the correct connection
initialization, so it seems reasonable to couple it with the connection
construction process.

Also updated FileConnection::connect() to be aligned with
DirectoryConnection::connect().

Change-Id: Ibf55761e1fdc4d8365ecddf7d8e52aa130056bde