[debugger] Merge NameLookupResult into FoundName.
Unifies name lookup into one function.

Adds the ability to look up an exact name in the index givena pre-split
identifier. This is important because the symbol system doesn't know how
to parse C++ (only "expr" does), so the splitting needs to be done at a
different level. The existing function is left currently to avoid
updating all callers at once.

A function is added on Identifier to producet the input format expected
by the symbol index.

Some initial work is done to look up template types. This is incomplete
and not yet tested. It will be filled out more in a followup.

Change-Id: I47b9079da36d726567f4456ac9563d8c8fd00231