[kernel][kcounter] Move kernel counters VMO format header out of system/private
This is private between kernel/lib/counters code and the isolated
low-level service that reads the VMOs.  Currently that's just the
simple test and uapp.  Eventually it will be some service thingy,
but still only a single a thing that can specially depend on the
kernel source this same way.

Bug: ZX-3845 #comment kernel counters VMO format header out of system/private
Change-Id: If0c963173c262a315af4d476116ed3459ab1196c