[fidlcat] Initial tracing code.
Initial support for tracing zx_channel_write calls and turning them into readable JSON.

This does not support all kinds of FIDL calls - in fact, it crashes on many of them.

The code in lib/ is pretty reasonable.  The driver program is in need of organization
and refactoring, but can probably be checked in as-is and iterated over.

Change-Id: I2a6593f1086802f846fc340dad9cc70103b93823