[build] Limit the number of concurrent jobs
When using large number of parallel jobs, e.g. -j1000 which is fairly
common when using goma, we would also spawn large number of actions that
could be memory intensive: Go build, rustc, Dart analyzer, as well as
linker. These are all competing for memory and could lead to
out-of-memory errors.

Previously, we used a fixed size pool for Dart analyzer to address this
issue, but that solution is also incorrect, because the the fixed size
is either going to be too large or too small depending on the amount of
memory the machine has. Rather, we should compute that number based on
the amount of available memory and the requirement of each type of job.

This change introduces new pools for linking, Rust, Go and Dart. The
size of these pools is computed using a custom script that's executed
using the exec_script function. The amount of memory requested for each
pool has been determined by performing local measurement on my
workstation and while they may not be perfectly accurate, they should be
a reasonable starting point (I've rounded up all the numbers to nearest
gigabyte to leave some extra room). We could fine tune these number
after we get more experience after using this change on bots and
developer workstations.

Note that this change doesn't restrict parallelism for other types of
actions we're running in our build, for example the FIDL generator. We
might consider restricting those as well if it turns out that they're
memory intensive.

Change-Id: I3fee60f4901743a0e7c25ae4c82d52f9a58537d8