[bt][tests] Allow fake hci naming and re-enable integration tests
Allow fake hci device names to be provided, to prevent the risk of name
clashes in tests using >1 fake hci-devices. Previously the random names
had only 8 bits of entropy, so 1/256 chance of a collision which
produced unexpected behavior and silent failures.

The LE.Central and Control integration tests were previous disabled due
to FLK-106; as that is now fixed we can re-enable these tests.

tests:
  $ fx run-tests bluetooth-tests -t bt-integration-tests

Change-Id: I2fd15e4b1facbb2fe6047cf94834f9780038f49d