[wayland][scanner] Fix panic decoding enum arguments.
The primitive type (ex: uint vs int) is annotated on the arguments for
an enum and not the enum itself. This means we need to agree on what
type we want to deserialize an enum value out of the message stream as.

This change simply always considers enum arguments to be uint values
on the wire. Transforming this into a signed value will handled by the
generated enum type.

PD-195 #done

Test: Launch weston-fullscreen, press 't' to set the buffer transform,
      observed no panic (and the surface rotates since we're not
      yet implementing set_buffer_transform).
Change-Id: I014bc018ce7ca6f851ef583406502173da7f554f