[debugger] Fix crash when root job is already attached.
This would happen when we want to attach to a new instance of the debug
agent. This new agent will attempt to bing to the root job, which is
already bound by the first.

There was a bug with the async handles that would make the async_loop
try to free a resource that wasn't acquired.

TEST=manual

Change-Id: Ibffaee312df16524e1ffc1af783776e273f4f48d