[debugger] Hook up type name finding.
Hook up all variants of name lookup to find types, templates, and
namespace names.

Adds tests for the new functionality for global name lookup as implicit
lookup on the "this" object.

Previously the FindName and variant functions returned an
std::optional<FoundName> but FoundName but FoundName encodes a "not
found" state itself. Changed these functions to return a FoundName
directly which makes the code a little cleaner.

Change-Id: I83a1a06ce926ee05ec1387f907ddce3be69f13bd