[docs][ftp] Import FTP-036 into markdown
Note that this includes a small change to the README.md
to note that we sort by review date, adds FTP-016, and
separates out the Accepted vs Rejected FTPs.

Finally, FTP-003 is updated as the URL for the //zircon...types.fidl
file changed, and this CL was going to trample a CL that fixed
that.

Change-Id: I8fdad8235a07fba343b3a32604b94ef06a40c686