[debugger] Track component controller/stdout.
The debug agent now keeps track of the components controller, which
enables it to kill it the correct way (through the component manager)
instead of killing the process directly.

We also obtain the stdout/stderr sockets and are able to pipe the data
out to the client.

Still a couple of details are missing:
- Correctly handle process exit (clear state).
- Have a better way to uniquely identify the running process.
- Better handle environments.

TEST=manual

Change-Id: I8f440fa8551841ecf8e8261f19be1de0809c04cf