[debugger] Merge --symbol-path and --symbol-repo-path
If we pass repo paths to --symbol-path we can detect whether there is a
.build-id folder present and treat it as a repo if so, and a simple
folder of objects if not. This should be adequate, so we can kill the
separate repo argument.

Along the way we no longer probe for the user's HOME folder to use as a
default symbol repo. This default isn't taken advantage of by anyone
today, and its original purpose was to serve as the default write
location once we have symbol downloading. It looks like writable repos
will be specified orthogonally anyway, so we can remove this for now.

Change-Id: I0b74d607f944b4461d7722031320d90d59f64190