[debugger] Hook up template casting.
Hooks up template matching to name lookup. This was mostly implementing
a TODO in global name lookup and factoring the code so it can be shared
by both implicit and exact name lookup.

Some prefix matching code has moved from the root ModuleSymbolIndex to
the ModuleSymbolIndexNode so that a prefix search can be done from
anywhere in the index tree.

Change-Id: I5b202a2633ef8ba3177817f7c5bdb13d21727305