[fx rust] fix discovery of FUCHSIA_BUILD_DIR
This should be wrapped in scripts that use fx-config-read, and so should
instead consume $FUCHSIA_BUILD_DIR.

Note that `out_dir` is a misnomer here, and should be renamed to build_dir. I
avoided changing the public flag surface as I have no way to assert if that
breaks other things.

Change-Id: Ie38fb874b9360e597deb6d5c28d14ba329e1cf8d