[ktl] Make ktl types true aliases
Previously, ktl was using templated using statements or simple function
wrappers to create aliases for std types, which is fine in some cases
but sometimes leaves out members of an overload set that could help with
optimization or allow different arguments.

In fact, in building with this I already caught a few misuses of
ktl::move since clang's std::move plugs into special compiler warnings
that the previous wrapper wasn't able to use.

Now we just use simple using declarations to import the std types we
want exactly as they're defined in the STL.

TEST: no functional change
Change-Id: Ib890527daec7b4847fc912df393e129d57425b71