[pkg] hex_serde should decode from a String
Before this patch, hex_serde::deserialize deserialized from
a `&str`, which allowed it to take advantage of serde's
zero copy support. Unfortunatately, not all deserializers
support zero copy, like serde_json::Value, and so it will
raise an error trying to deserialize into a `&str`. Since
this isn't on a hotpath, the easiest thing to do is
just to deserialize into a String.

Change-Id: Ibe16ccd656f85493775553358c12703b52da7d49